+++ Content/Items/Weapons/Summon/SafetyLamp/SafetyLampMinion.cs
@@ -100,7 +100,12 @@ namespace SuperfluityTwo.Content.Items.Weapons.Summon.SafetyLamp
 		{
 			foreach (var npc in Main.ActiveNPCs)
 			{
-				if (npc.CanBeChasedBy() &&
+				if (npc.active &&
+					!npc.CountsAsACritter &&
+					!npc.dontTakeDamage &&
+					!npc.friendly &&
+					!npc.immortal &&
+					Main.player[Projectile.owner].CanNPCBeHitByPlayerOrPlayerProjectile(npc) &&
 					Collision.CanHit(Projectile.Center, 0, 0, npc.position, npc.width, npc.height) &&
 					npc.Hitbox.ClosestPointInRect(Projectile.Center).Distance(Projectile.Center) < Range)
 				{